[Core] Minor change in `FileSerializer`
